Backend/requirements.txt
# Django and Web Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
pymongo==4.6.0
djongo==1.3.6

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7

# Blockchain Integration
web3==6.11.3
eth-account==0.9.0

# Biometric & AI
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3

# File Upload & Media
Pillow==10.1.0
django-storages==1.14.2

# API Documentation
drf-yasg==1.21.7

# Environment Management
python-decouple==3.8

# Utilities
requests==2.31.0
celery==5.3.4
redis==5.0.1

# Development Tools
django-extensions==3.2.3
